Add run_in_terminal option to disable cooked mode
Adds a `cooked_mode` optional argument to run_in_terminal. Using it
avoids CPR responses being echoed to the terminal (see #482). It is not
a complete solution because they can still appear once the prompt exits
and thus leaves raw mode, but avoids some instances.

The default is left at the old behaviour for backwards compatibility.