Set the strings attribute on DummyHistory
Several places in the code assume history objects have the strings attribute
(and that it's a list).

We could add strings as an abstractproperty on the abc, but then every history
class would have to define strings as a property, which would generally require
using a getter and a setter. I'm not sure if we want that level of
complication, so I haven't done it here.