Fixed WordCompleter when line ends with a space
This was broken by commit 81c30157 (add support for patterns/fuzzy word completer).
Completing 'test ' no longer returned any options.

Added unittests to expose the issue.

Also added unittests to verify the FuzzyCompleter supports lines ending with a space
(which was the case but at least now it should be impossible to break that in the future)