Make sure HANDLE on Windows has a correct size
Previously, the type of various HANDLEs are native Python integer types.
The ctypes library will treat them as 4-byte integer when used in function
arguments. On 64-bit Windows, HANDLE is 8-byte and usually a small integer.
Depending on whether the extra 4 bytes are zero-ed out or not, things can
happen to work, or break.

This patch adds explicit type casts so ctypes uses 8-byte integers for
HANDLEs on 64-bit Windows.

Thanks to @quark-zju for providing this patch to the 1.0 branch. This patch
contains similar changes for the the 3.0 branch.