Allow empty continuation.
If continuation is an empty string, instead of ignoring it and aligning with
spaces, honor the non-None but empty value and don't align to the prompt width
with spaces.

This is helpful if you want to configure your prompt to allow multiline editing
and also want to copy the text directly from your terminal.

Commit edited by Jonathan Slenders.