Use a win32 Event to notify the loop in `_Win32Handles` to stop waiting for stdin.
This fixes a bug where the prompt_toolkit application is stopped by something
other than user input. (E.g., a progress bar that terminates.) In that case,
the thread which handles the user input will keep running and prevent Python
from exiting.

In the case of a progress bar, it was even worse, because there we close the
asyncio event loop when we finish (because it's a new loop, running in a
different thread). And closing the loop causes the `call_soon_threadsafe` in
the Win32 handles to raise a `RuntimeError` ("RuntimeError: Event loop is
closed").