Fix race condition in `ProgressBar` shutdown logic
The app might not have started upon `__exit__`ing
the context manager from the main thread;
in which case `self.app.exit()` has no effect,
the app will be left to run indefinitely
and the program will hang after `join`ing
the app thread.
Instead, we can repurpose `on_invalidate`
to call `app.exit` the next time
the app is invalidated.
This relies on the app's `refresh_interval`
being set to a value other than `None` for it
to happen automatically.

Closes #964.