Guarantee height of at least 1 for both labels and text areas.
This solves an issue where we have a `VSplit`, consisting of many labels, with
a precise padding value given. If the window is too small, then the padding
will take up all the space (because the dimension is "exact"), and no space
will remain for the labels (where the minimum size given was 0 instead of 1).

It doesn't make sense for labels and text areas to have a zero size, so we
increase this to 1.