Fix bug preventing documentation search from working in some cases.
In some cases, the documentation search does not work. This issue
was introduced in 279720474de378d6a970efc0858b595d798c5163 when
Sphinx was upgraded. The relevant issue is:

 - https://github.com/sphinx-doc/sphinx/issues/8623

For some reason, the `sphinx_rtd_theme` theme used for this project
is constrained to `<0.5` (not sure where this has been set). This PR
forces the version of `sphinx_rtd_theme` to a working version.