Stop preferring a TTY output when creating the default output.
Use PlainTextOutput instead if `sys.stdout` is not a TTY. Especially for
printing formatted text, it is useful for the output to be redirected to a
file. In that case, we don't want to include ANSI escape sequences, but have a
plain text output. Similar to how we can read input from something that was
piped into stdin.