Fix control-c key binding for progressbar when created in other thread + cancel_callback.
If the ProgressBar was created in a thread that is not the main thread, the
control-c key binding should not do anything. It was killing the wrong thread.

On top of this, accept a `cancel_callback` that will be called when control-c
is pressed. This can be used to start cancellation of the code that the
progress bar is wrapping.