Make VERSION tuple numeric.
Note that by being numeric we drop the rc/a/b/dev suffix.
I'm also adding the pep440 regular expression to make sure the version
number match the spec.

The pep 440 spec is a bit annoying with the fact that a/b/rc don't have
dots, but .dev do.

There are some issues if you don't follow the spec and publish both whl/tgz, as
pip will mix-up the values, and think the tgz are actually a more recent
version than the whl. It may be fixed in setuptools/pip, but better be
safe.

Ii did not go the route of setting VERSION and having __version__ be a
'.'.join() as setup.py use regex to extract version number.

Maye docs/conf.py could also use the same trick as setup.py to avoid
having to update the version number in multiple places ?

Closes #1707