Fix line wrapping in `patch_stdout` on Windows.
Enable autowrap before writing data to the output in `patch_stdout`. We
enable/disable VT terminal processing during every `flush()` of the Windows
Output implementation, but it turns out that this causes the autowrap flag to
be reset each time. Because of this, we have to enable autowrap each time
before writing.