Refactoring of StdoutProxy to take app_session from the caller
This changes makes StdoutProxy to take the exact app_session from
by taking it as optional argument in the constructor instead finding
it by calling get_app_session() globally.

It can avoid confliction of printing output within the patch_stdout
context in the ssh session when the multiple ssh connections are
performing concurrently.

The changed StdoutProxy now passes the correct app instance from the
given app_session in the constructor to the run_in_terminal() in it
instead of calling get_app_or_none() globally that can give wrong
app instance from the prompt session in the last ssh connection.