Minimal changes to allow multiline suggestions.
I'm trying to be the least invasive;
But the idea is to allow access to previous lines when rendering a line
in order to shift them down; and display multiline sugestions.

With this I can add minimal modification into IPython that allows
multiline suggestion (typically would be created by an LLM – while not a
big fan, there is a lot of requests for it.

I'm going to assume there is no willingness tochange the current API
for transformers to get at least a ranges of lines instead of lines,
it would be easier to handle.

I do not include any update to the transformers we use in IPython to
display those completions; but once battled tested a bit more, I'm happy
to contribute it as well.