Try to run ruff and typo only on 3.13
There is no need to run N times.

This in particular means that we can still see from a glace the tests
are passing on older python.

I also try to run typo regardless of wether ruff check and formating
passes, which hopefully avoid one iteration of fix, push and CI failing.