Add ability to change the autosuggestion class.
Mostly up for discussion;but while working with multiline completion and
LLM, I realised that in IPython it is likely Ii also want to handle
suggestion when on the last line, so I think i'd like to ed the
autocompltion class to a no-op one.

Ideally I think it woudl make sens to completly control the full
conditional processor; so mostly openeing that for discussion.